Add network connectivity test
also add a skip-broken flag to yum update for the time-being
because a syslinux dependency: syslinux-nonlinux was not found
in centos mirrors. It is on the way though.
https://www.centos.org/forums/viewtopic.php?f=14&t=45933
Need 'yum clean all' to update metadata and get the latest package
list from mirros.
Fix a bug in prepare.sh. it didn't fetch the gerrit refs.
Restore some package pre-load scripts in prepare.sh to make CI faster.

Change-Id: I66f06bdfb801204fb956c7a264c98de879322744