Increase disk size to avoid insufficient space for tempest
Sometimes when running volume related tempest tests, there is
change that such error occurs: "Failed to create ALL necessary
volumes to run test". It might because a race in deleting and
creating volumes. Anyway, 5G cinder volume is enough but not by
a margin. Give it a little more space.

Change-Id: Id129c21461fd027e4eddf9ea7aee24f74154dee7