Restore some config to increae cache hit
since the nature of packages is static given the particular version
number, it is safe to ignore packages cache header and use a larger
expiration time.
also enable access_log for monitoring the requests and hit rate so
as to tune the cache control configurations later based on that knowledge.

Change-Id: I8403ecca663ecc84facce7a215efebb553d225fc