Use download speed to decide which url is faster
just getting the root path of a http server to measure the response
time does not necessarily reflect the real download speed of a source.
reponse can be 302 http redirect which may be really fast but download
bandwidth is determined by the route to the real source http server.
Therefore, use download time for the first chunk of 20000 bytes to
determine which is faster.

Closes-bug: #1392816
Change-Id: I3019dce3cf92dd75c58a21987c653cbcb502264b