unpin versions of python-daemon, paramiko, etc
python-daemon 2.0.4 has fixed issue related to setuptools, so revert
the pin but excludes broken versions.
since we have moved compass-core, tempest into virtualenv, it is also not
necessary to pin other package versions to be compatbile with tempest.
Pin MarkDown to <2.5 because py26 is not supported otherwise.
python-daemon also needs a new pip that comes with new virtualenv.

Change-Id: I7f378f344ef2e232c10e77725202acb09b3c3393