Seperate chef installer as a plugin Bug: 1506700
This is the first change list of a series of changes that will eventually
convert chef installer as a plugin.
It moves chef installer related configure files to the plugins/chef_installer/
util.py's load_configs is modified to read not only the ordinary configs but
also the configs in each of plugins directory.

Change-Id: I285c59556f4abd89255608e15117e6d012ecfb01