Seperate chef installer as a plugin (move chef templates to plugins directory) Bug: 1506700
This is the second change list of a series of changes that will eventually
convert chef installer as a plugin.
It moves chef installer related templates files to the plugins/chef_installer/templats
PLUGINS_DIR was introduced to indicate where the plugins directory will be, it may be in /etc/compass or in a different place util.py's load_configs will use it when reading plugins configs.

Change-Id: Ieb599e3e7c8cecd8355e77d89e3da8548f78a93a